+++ src/analysis/assembler/operators/integration.h
@@ -8,12 +8,12 @@
 
 namespace espreso {
 
-template <int nodes, int gps, int ndim, int edim, int etype, class Physics> struct Integration;
+template <size_t nodes, size_t gps, size_t ndim, size_t edim, size_t etype, class Physics> struct Integration;
 
-template <int nodes, int gps, int edim, int etype, class Physics>
+template <size_t nodes, size_t gps, size_t edim, size_t etype, class Physics>
 struct Integration<nodes, gps, 2, edim, etype, Physics>: ActionOperator, Physics {
 
-	Integration(int interval) { isconst = false; }
+	Integration(size_t interval) { isconst = false; }
 
 	void sisd(typename Physics::Element &element)
 	{
@@ -71,10 +71,10 @@ struct Integration<nodes, gps, 2, edim, etype, Physics>: ActionOperator, Physics
 	}
 };
 
-template <int nodes, int gps, int edim, int etype, class Physics>
+template <size_t nodes, size_t gps, size_t edim, size_t etype, class Physics>
 struct Integration<nodes, gps, 3, edim, etype, Physics>: ActionOperator, Physics {
 
-	Integration(int interval) { isconst = false; }
+	Integration(size_t interval) { isconst = false; }
 
 	void sisd(typename Physics::Element &element)
 	{
@@ -167,10 +167,10 @@ struct Integration<nodes, gps, 3, edim, etype, Physics>: ActionOperator, Physics
 	}
 };
 
-template <int nodes, int gps, int ndim, int edim, int etype, class Physics>
+template <size_t nodes, size_t gps, size_t ndim, size_t edim, size_t etype, class Physics>
 struct Volume: ActionOperator, Physics {
 	double &volume;
-	Volume(int interval, std::vector<double> &volume): volume(volume[interval]) { isconst = false; }
+	Volume(size_t interval, std::vector<double> &volume): volume(volume[interval]) { isconst = false; }
 
 	void simd(typename Physics::Element &element)
 	{
